refactor folder structures
