+++ frontend/app/components/styles/SearchBar.module.css
